Move filename encoder to new `encoding` mod
